+++ farben-build/src/core.rs
@@ -1,0 +1,5 @@
+//! Exported library internals for users that may use them
+//!
+//! Includes parser stuff
+
+pub use crate::parser::*;